requirements.txt
streamlit
sentence-transformers
langchain
langchain-community
chromadb
python-dotenv
requests
pypdf
python-docx